refactor: PR comment changes
