+++ src/components/SSO/index.ts
@@ -1,0 +1,3 @@
+import SocialRegistration from './SSO'
+
+export default SocialRegistration
