--- src/components/SocialSSO/index.ts
@@ -1,3 +1,0 @@
-import SocialRegistration from './SocialSSO'
-
-export default SocialRegistration
